+++ elastic-job-lite/elastic-job-lite-spring/src/test/java/com/dangdang/ddframe/AllLiteSpringTests.java
@@ -24,5 +24,5 @@ import org.junit.runners.Suite.SuiteClasses;
 
 @RunWith(Suite.class)
 @SuiteClasses(AllSpringJobTests.class)
-public class AllLiteSpringTests {
+public final class AllLiteSpringTests {
 }